+++ apps/backend/src/progress/activeReviewDays/activeReviewDays.ts
@@ -1,8 +1,8 @@
-import type { DatabaseExecutor } from "../database";
+import type { DatabaseExecutor } from "../../database";
 import {
   formatDateAsTimeZoneLocalDate,
   requireIanaTimeZone,
-} from "./timeZone";
+} from "../timeZone";
 
 export type ReviewTimeZoneSource = "client" | "user_settings";
