+++ apps/backend/src/progress/activeReviewDays/activeReviewDaysBackfill.ts
@@ -3,15 +3,15 @@ import {
   applyWorkspaceDatabaseScopeInExecutor,
   type DatabaseExecutor,
   type SqlValue,
-} from "../database";
-import { withTransientDatabaseRetry } from "../database/transient";
-import { unsafeRepeatableReadTransaction, unsafeTransaction } from "../database/unsafe";
-import { withReportingReadOnlyTransaction } from "../admin/reportingDb";
+} from "../../database";
+import { withTransientDatabaseRetry } from "../../database/transient";
+import { unsafeRepeatableReadTransaction, unsafeTransaction } from "../../database/unsafe";
+import { withReportingReadOnlyTransaction } from "../../admin/reportingDb";
 import {
   captureBackendWarning,
   normalizeCaughtError,
   type BackendObservationScope,
-} from "../observability/sentry";
+} from "../../observability/sentry";
 import {
   materializeMissingActiveReviewDaysForUserInExecutor,
   type ActiveReviewDayMaterializationResult,
